AMP-1680,AMP-1696
git-svn-id: file:///svn/trunk@4316 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/aim/util/AmpMath.java
@@ -15,11 +15,11 @@ public class AmpMath {
     static byte PREC_UNARY = 12; // Not actually used.
     static byte PREC_NONE = 13;
 
-    public static double CalcExp(String exp) {
+    public static long CalcExp(String exp) {
         String expr = exp.toLowerCase();
         expr = expr.replace(" ", "");
         expr = expr.replace(",",".");
-        return CalcExp_Ex(expr);
+        return (long)CalcExp_Ex(expr);
     }
 
     private static double CalcExp_Ex(String exp) {
